+++ wrapper/fs_wrapper_fatpetit.h
@@ -7,6 +7,6 @@ fswresult_e fs_wrapper_fatpetit_opendir(fs_wrapper_dir_t *dir, const char *dirpa
 fswresult_e fs_wrapper_fatpetit_readdir(fs_wrapper_dir_t *dir, fs_wrapper_diritem_t *item);
 fswresult_e fs_wrapper_fatpetit_open(const char *fname);
 fswresult_e fs_wrapper_fatpetit_read(void *readbuf, uint16_t size);
-fswresult_e fs_wrapper_fatpetit_seek(uint16_t offset);
+fswresult_e fs_wrapper_fatpetit_seek(uint32_t offset);
 
 #endif /*FS_WRAPPER_FATPETIT_H_*/
