+++ wrapper/fs_wrapper_linux.h
@@ -9,6 +9,6 @@ fswresult_e fs_wrapper_linux_opendir(fs_wrapper_dir_t *dir, const char *dirpath)
 fswresult_e fs_wrapper_linux_readdir(fs_wrapper_dir_t *dir, fs_wrapper_diritem_t *item);
 fswresult_e fs_wrapper_linux_open(const char *fname);
 fswresult_e fs_wrapper_linux_read(void *readbuf, uint16_t size);
-fswresult_e fs_wrapper_linux_seek(uint16_t offset);
+fswresult_e fs_wrapper_linux_seek(uint32_t offset);
 
 #endif /*FS_WRAPPER_LINUX_H_*/
